Bug fix for Fast5_to_seq_summary for directories with 1 fast5 file
Currently, the code looks at the enumerate index `i` to see if a fast5 file was found in the given directory.
However, if a directory does not have any files in it, then the enumeration loop won't get started and the
`i` variable will actually be unbound, so the code as written would raise an `UnboundLocalError`.
If there is a single fast5 file in the directory, then `i` will be valid, but it will be `==0`.

I've updated the logic to use a simple flag to indicate when we have actually inserted a fast5 file
into the Queue during this loop.